stockexchange/requirements.txt
appdirs==1.4.4
asgiref==3.3.4
authy==2.2.6
awsebcli==3.20.2
bcrypt==3.2.0
beautifulsoup4==4.9.3
botocore==1.21.65
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.3
cryptography==36.0.0
decorator==4.4.2
distlib==0.3.1
Django==3.2
django-chartjs==2.2.1
django-crispy-forms==1.11.2
django-qrcode==0.3
django-rest-knox==4.1.0
django-scss==0.6.1
djangorestframework==3.12.4
docutils==0.17.1
filelock==3.0.12
Flask==1.1.2
fusionexport==1.2.2.1
future==0.16.0
glob2==0.7
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jmespath==0.10.0
lxml==4.6.3
MarkupSafe==1.1.1
multitasking==0.0.10
numpy==1.21.4
pandas==1.3.4
pathspec==0.5.9
Pillow==8.2.0
pipenv==2020.11.15
plivo==4.10.0
pycparser==2.20
PyJWT==1.7.1
pyotp==2.6.0
pypiwin32==223
python-dateutil==2.8.2
pytz==2021.1
pywin32==302
PyYAML==5.4.1
razorpay==1.2.0
requests==2.25.1
segno==1.3.3
semantic-version==2.8.5
shell==1.0.1
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
statistics==1.0.3.5
termcolor==1.1.0
twilio==6.58.0
urllib3==1.26.7
virtualenv==20.4.2
virtualenv-clone==0.5.4
virtualenvwrapper-win==1.2.6
wcwidth==0.1.9
Werkzeug==1.0.1
yfinance==0.1.67
